neo4j-mdr-db/utils/activities/requirements.txt
pyyaml
neo4j
stringcase
python-dotenv
requests
py2neo
service
fastapi
pydantic
pytest
uvicorn
pandas
numpy
openpyxl
datetime